chore: rearranged component files
