+++ components/index.ts
@@ -6,3 +6,5 @@ export { CodeBlock } from './CodeBlock'
 export { InlineCode } from './InlineCode'
 export { Link } from './Link'
 export { PropsTable } from './PropsTable'
+export { Main } from './Main'
+export { Navigation } from './Navigation'
